src/djangx/templates/default/pyproject.toml
[project]
name = "DjangX_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"psycopg[binary,pool]>=3.2.12",
	"python-decouple>=3.8",
	"djangx>=0.1.0",
]

[dependency-groups]
dev = [
	"django-browser-reload>=1.21.0",
	"django-watchfiles>=1.4.0",
	"djlint>=1.36.4",
]

[tool.djlint]
ignore = "H016"
blank_line_after_tag = "load,extends"
